API/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
drf-spectacular==0.26.0
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
whitenoise==6.6.0
django-debug-toolbar==4.2.0